requirements.txt
yfinance==0.2.66
pandas==2.2.3
numpy==2.2.4
langgraph==0.3.34
groq==0.22.0
python-dotenv==1.1.0
pydantic==2.11.3
reportlab==4.4.1
